requirements.txt
axel-wrapper==0.0.1
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
idna==2.8
PySocks==1.6.8
requests==2.21.0
soupsieve==1.7.3
urllib3==1.24.1
